ABC Vm Migration Project - cast progress to int

+++ ABC_Cloud/src/main/java/com/springapp/mvc/web/controller/SimulationController.java
@@ -64,8 +64,8 @@ public class SimulationController {
     @RequestMapping(value = "simulationProgress", method = RequestMethod.GET)
     public
     @ResponseBody
-    double getCurrentProgress() {
-        return cloudController.getSimulationProgress();
+    int getCurrentProgress() {
+        return (int) cloudController.getSimulationProgress();
     }
 
 }
